requirements.txt
Flask==2.3.2
pymongo==4.3.3
python-dotenv==1.0.0